zspaimai_ui/requirements.txt
pytest
pyyaml
request
